src/API/requirements.txt
flask
tensorflow
keras
requests
sqlalchemy
psycopg2-binary
pybase64
flask_cors
pillow
opencv-python
pytesseract
numpy